ml-service/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6
# pdfplumber==0.10.3
# python-docx==1.1.0
# spacy==3.8.2
# nltk==3.8.1
# numpy>=2.0.0
# python-jose[cryptography]==3.3.0
# pydantic==2.5.0
# aiofiles==23.2.1

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
aiofiles==23.2.1

# Document Parsing
pdfplumber==0.10.3
python-docx==1.1.0

# NLP & NER
spacy==3.7.2
transformers==4.36.0
torch==2.2.0

# Skill Extraction & Matching
sentence-transformers==2.3.0
scikit-learn==1.3.2

# ML Models
xgboost==2.0.2
numpy==1.26.2
pandas==2.1.3

# Feature Engineering
dateparser==1.2.0
regex==2023.10.3

# Utilities
python-jose[cryptography]==3.3.0
huggingface-hub==0.20.0
